refactor(gatsby-plugin-manifest): don't import fs in ssr code
