+++ packages/gatsby-plugin-manifest/src/__tests__/common.js
@@ -1,22 +1,9 @@
-const path = require(`path`)
-const { defaultIcons, doesIconExist, addDigestToPath } = require(`../common`)
+const { defaultIcons, addDigestToPath } = require(`../common`)
 
 describe(`gatsby-plugin-manifest`, () => {
   describe(`defaultIcons`, () => {
     it(`includes all icon sizes`, () => {
       expect(defaultIcons).toMatchSnapshot()
-    })
-  })
-
-  describe(`doesIconExist`, () => {
-    it(`returns true if image exists on filesystem`, () => {
-      const iconSrc = path.resolve(__dirname, `./images/gatsby-logo.png`)
-      expect(doesIconExist(iconSrc)).toBeTruthy()
-    })
-
-    it(`returns false if image does not exist on filesystem`, () => {
-      const iconSrc = path.resolve(__dirname, `./images/non-existent-logo.png`)
-      expect(doesIconExist(iconSrc)).toBeFalsy()
     })
   })
